Added depth limit
description: Added depth limit parameter to the alphabeta/6 predicate to stop the AI from generating too many states in the state graph tree and not respond.

The higher the depth the smarter the AI and the longer it takes to respond.

And made the proper adjustments to the alphabeta algorithm to match this adjustment.

Also made the necessary adjustments to the driver predicate play/2.